Added Relative coordinate option
When the relative clicking mode's on, the coordinates collected are x and y differences to the last coordinate instead of just raw coordinates. Now the clicks are heavily based on your current mouse position.
The first click that's put in the queue won't be put in the queue, but the coordinates will be used for the next click's relative coordinates. To remove the first click: empty the entire array with the pop or empty hotkeys.
(For now, change the "rel" variable from 0 to 1)